Move classes to specified directories.
