+++ src/dev/classes/application/Settings.py
