+++ src/dev/classes/updates/UpdateChecker.py
@@ -1,4 +1,4 @@
-from src.dev.classes.Addon import Addon
+from src.dev.classes.addons.Addon import Addon
 import logging
