+++ src/window_classes/AddonWindow.py
@@ -3,9 +3,9 @@ from PyQt5.QtWidgets import QDialog, QMessageBox
 from PyQt5.QtCore import pyqtSlot, pyqtSignal
 import requests
 import requests.exceptions
-from src.dev.classes.Addon import Addon
+from src.dev.classes.addons.Addon import Addon
 import logging
-from src.dev.classes.Worker import Worker
+from src.dev.classes.workers.Worker import Worker
 
 supported_sites = ['curse-projects', 'curse-addons', 'tukui']
